+++ apps/api/plane/app/views/module/issue.py
@@ -36,7 +36,8 @@ from plane.utils.grouper import (
 from plane.utils.issue_filters import issue_filters
 from plane.utils.order_queryset import order_issue_queryset
 from plane.utils.paginator import GroupedOffsetPaginator, SubGroupedOffsetPaginator
-from plane.utils.filters import ComplexFilterBackend
+# Evolury: backend com propriedade personalizada (ADR 0011)
+from plane.utils.filters import FiltroComPropriedades
 from plane.utils.filters import IssueFilterSet
 from .. import BaseViewSet
 from plane.utils.host import base_host
@@ -47,7 +48,7 @@ class ModuleIssueViewSet(BaseViewSet):
     model = ModuleIssue
     webhook_event = "module_issue"
     bulk = True
-    filter_backends = (ComplexFilterBackend,)
+    filter_backends = (FiltroComPropriedades,)
     filterset_class = IssueFilterSet
 
     def apply_annotations(self, issues):